requirements.txt
asgiref==3.5.2
Django==4.1
django-ckeditor==6.5.0
django-crispy-forms==1.9.2
django-js-asset==2.0.0
django-widget-tweaks==1.4.12
Pillow==9.2.0
sqlparse==0.4.2
tzdata==2022.2
whitenoise==6.2.0
waitress==2.1.2